organize settings file structure
