+++ src/user-federation/UserFederationLdapSettings.tsx
@@ -2,13 +2,13 @@ import { PageSection } from "@patternfly/react-core";
 import { useTranslation } from "react-i18next";
 import React from "react";
 import { ScrollForm } from "../components/scroll-form/ScrollForm";
-import { LdapSettingsAdvanced } from "./LdapSettingsAdvanced";
-import { LdapSettingsKerberosIntegration } from "./LdapSettingsKerberosIntegration";
-import { LdapSettingsCache } from "./LdapSettingsCache";
-import { LdapSettingsSynchronization } from "./LdapSettingsSynchronization";
-import { LdapSettingsGeneral } from "./LdapSettingsGeneral";
-import { LdapSettingsConnection } from "./LdapSettingsConnection";
-import { LdapSettingsSearching } from "./LdapSettingsSearching";
+import { LdapSettingsAdvanced } from "./ldap/LdapSettingsAdvanced";
+import { LdapSettingsKerberosIntegration } from "./ldap/LdapSettingsKerberosIntegration";
+import { LdapSettingsCache } from "./ldap/LdapSettingsCache";
+import { LdapSettingsSynchronization } from "./ldap/LdapSettingsSynchronization";
+import { LdapSettingsGeneral } from "./ldap/LdapSettingsGeneral";
+import { LdapSettingsConnection } from "./ldap/LdapSettingsConnection";
+import { LdapSettingsSearching } from "./ldap/LdapSettingsSearching";
 
 export const UserFederationLdapSettings = () => {
   const { t } = useTranslation("user-federation");
