+++ src/user-federation/kerberos/KerberosSettingsRequired.tsx
@@ -8,13 +8,13 @@ import {
   TextInput,
 } from "@patternfly/react-core";
 import { useTranslation } from "react-i18next";
-import { HelpItem } from "../components/help-enabler/HelpItem";
+import { HelpItem } from "../../components/help-enabler/HelpItem";
 import { useForm, Controller } from "react-hook-form";
 import ComponentRepresentation from "keycloak-admin/lib/defs/componentRepresentation";
-import { FormAccess } from "../components/form-access/FormAccess";
-import { useAdminClient } from "../context/auth/AdminClient";
+import { FormAccess } from "../../components/form-access/FormAccess";
+import { useAdminClient } from "../../context/auth/AdminClient";
 import { useParams } from "react-router-dom";
-import { convertToFormValues } from "../util";
+import { convertToFormValues } from "../../util";
 
 export const KerberosSettingsRequired = () => {
   const { t } = useTranslation("user-federation");
