+++ src/user-federation/ldap/LdapSettingsCache.tsx
@@ -6,67 +6,20 @@ import {
   TextInput,
 } from "@patternfly/react-core";
 import { useTranslation } from "react-i18next";
-import { HelpItem } from "../components/help-enabler/HelpItem";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
+import { HelpItem } from "../../components/help-enabler/HelpItem";
 import { useForm, Controller } from "react-hook-form";
-import { convertToFormValues } from "../util";
 import ComponentRepresentation from "keycloak-admin/lib/defs/componentRepresentation";
-import { FormAccess } from "../components/form-access/FormAccess";
-import { useAdminClient } from "../context/auth/AdminClient";
-import { useParams } from "react-router-dom";
+import { FormAccess } from "../../components/form-access/FormAccess";
 
-export const KerberosSettingsCache = () => {
+export const LdapSettingsCache = () => {
   const { t } = useTranslation("user-federation");
   const helpText = useTranslation("user-federation-help").t;
-
-  const adminClient = useAdminClient();
-  const { register, control, setValue } = useForm<ComponentRepresentation>();
-  const { id } = useParams<{ id: string }>();
-
-  const convertToDays = (num: string) => {
-    switch (num) {
-      case "1":
-        return t("common:Sunday");
-      case "2":
-        return t("common:Monday");
-      case "3":
-        return t("common:Tuesday");
-      case "4":
-        return t("common:Wednesday");
-      case "5":
-        return t("common:Thursday");
-      case "6":
-        return t("common:Friday");
-      case "7":
-        return t("common:Saturday");
-      default:
-        return t("common:selectOne");
-    }
-  };
-
-  const setupForm = (component: ComponentRepresentation) => {
-    Object.entries(component).map((entry) => {
-      if (entry[0] === "config") {
-        convertToFormValues(entry[1], "config", setValue);
-        setValue("config.evictionDay", convertToDays(entry[1].evictionDay[0]));
-      } else {
-        setValue(entry[0], entry[1]);
-      }
-    });
-  };
-
-  useEffect(() => {
-    (async () => {
-      const fetchedComponent = await adminClient.components.findOne({ id });
-      if (fetchedComponent) {
-        setupForm(fetchedComponent);
-      }
-    })();
-  }, []);
 
   const [isCachePolicyDropdownOpen, setIsCachePolicyDropdownOpen] = useState(
     false
   );
+
   const [isEvictionHourDropdownOpen, setIsEvictionHourDropdownOpen] = useState(
     false
   );
@@ -74,6 +27,7 @@ export const KerberosSettingsCache = () => {
     isEvictionMinuteDropdownOpen,
     setIsEvictionMinuteDropdownOpen,
   ] = useState(false);
+
   const [isEvictionDayDropdownOpen, setIsEvictionDayDropdownOpen] = useState(
     false
   );
@@ -92,6 +46,8 @@ export const KerberosSettingsCache = () => {
     minuteOptions.push(<SelectOption key={index + 1} value={index} />);
   }
 
+  const { control, register } = useForm<ComponentRepresentation>();
+
   return (
     <>
       {/* Cache settings */}
@@ -108,7 +64,7 @@ export const KerberosSettingsCache = () => {
           fieldId="kc-cache-policy"
         >
           <Controller
-            name="config.cachePolicy"
+            name="cachePolicy"
             defaultValue=""
             control={control}
             render={({ onChange, value }) => (
@@ -125,12 +81,10 @@ export const KerberosSettingsCache = () => {
                 }}
                 selections={value}
                 variant={SelectVariant.single}
+                // aria-label="Other"
+                // isDisabled
               >
-                <SelectOption
-                  key={0}
-                  value={t("common:selectOne")}
-                  isPlaceholder
-                />
+                <SelectOption key={0} value="Choose..." isPlaceholder />
                 <SelectOption key={1} value="DEFAULT" />
                 <SelectOption key={2} value="EVICT_DAILY" />
                 <SelectOption key={3} value="EVICT_WEEKLY" />
@@ -140,6 +94,8 @@ export const KerberosSettingsCache = () => {
             )}
           ></Controller>
         </FormGroup>
+
+        {/* TODO: Field shows only if cache policy is EVICT_WEEKLY */}
         <FormGroup
           label={t("evictionDay")}
           labelIcon={
@@ -152,7 +108,7 @@ export const KerberosSettingsCache = () => {
           fieldId="kc-eviction-day"
         >
           <Controller
-            name="config.evictionDay"
+            name="evictionDay"
             defaultValue=""
             control={control}
             render={({ onChange, value }) => (
@@ -186,6 +142,8 @@ export const KerberosSettingsCache = () => {
             )}
           ></Controller>
         </FormGroup>
+
+        {/* TODO: Field shows only if cache policy is EVICT_WEEKLY or EVICT_DAILY */}
         {/* TODO: Investigate whether this should be a number field instead of a dropdown/text field */}
         <FormGroup
           label={t("evictionHour")}
@@ -199,7 +157,7 @@ export const KerberosSettingsCache = () => {
           fieldId="kc-eviction-hour"
         >
           <Controller
-            name="config.evictionHour"
+            name="evictionHour"
             defaultValue=""
             control={control}
             render={({ onChange, value }) => (
@@ -221,6 +179,8 @@ export const KerberosSettingsCache = () => {
             )}
           ></Controller>
         </FormGroup>
+
+        {/* TODO: Field shows only if cache policy is EVICT_WEEKLY or EVICT_DAILY */}
         {/* TODO: Investigate whether this should be a number field instead of a dropdown/text field */}
         <FormGroup
           label={t("evictionMinute")}
@@ -234,7 +194,7 @@ export const KerberosSettingsCache = () => {
           fieldId="kc-eviction-minute"
         >
           <Controller
-            name="config.evictionMinute"
+            name="evictionMinute"
             defaultValue=""
             control={control}
             render={({ onChange, value }) => (
@@ -256,6 +216,8 @@ export const KerberosSettingsCache = () => {
             )}
           ></Controller>
         </FormGroup>
+
+        {/* TODO: Field shows only if cache policy is MAX_LIFESPAN */}
         <FormGroup
           label={t("maxLifespan")}
           labelIcon={
@@ -271,7 +233,7 @@ export const KerberosSettingsCache = () => {
             isRequired
             type="text"
             id="kc-max-lifespan"
-            name="config.maxLifespan"
+            name="maxLifespan"
             ref={register}
           />
         </FormGroup>
